添加Swagger2实现在线api文档

+++ src/main/java/com/yoha/mall/mbg/mapper/PmsBrandMapper.java
@@ -2,11 +2,12 @@ package com.yoha.mall.mbg.mapper;
 
 import com.yoha.mall.mbg.model.PmsBrand;
 import com.yoha.mall.mbg.model.PmsBrandExample;
-import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
+import java.util.List;
+
 public interface PmsBrandMapper {
-    int countByExample(PmsBrandExample example);
+    long countByExample(PmsBrandExample example);
 
     int deleteByExample(PmsBrandExample example);
